requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
datasets==2.19.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
logzero==1.7.0
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
openai==1.30.1
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ragrank==0.0.7
requests==2.32.2
requests-toolbelt==1.0.0
six==1.16.0
smartapi-python==1.4.8
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
pre-commit
celery==5.2.7
redis==5.0.4
flower
pyotp
billiard==3.6.4.0
torch
ultralytics
transformers
mplfinance
huggingface_hub[hf_xet]
scikit-learn
joblib
python-dotenv
matplotlib
tqdm
scipy
scikit-image
scikit-learn
trl
stable-baselines3
gymnasium
